# apidog-import-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Feedbase Backend API (APIdog Import)",
    "version": "1.0.0",
    "description": "APIdog import file generated from API_FULL_LIST.md."
  },
  "servers": [
    {
      "url": "http://localhost:4560",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/notifications/list": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "POST /notifications/list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Notifications retrieved successfully",
                  "data": {
                    "notifications": [
                      {
                        "id": 901,
                        "notification_type": "post_status",
                        "title": "Post moved to planned",
                        "message": "Add dark mode is now planned.",
                        "is_read": 0
                      }
                    ],
                    "total": 4
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "paginationData": {
                    "type": "object",
                    "properties": {
                      "itemsPerPage": {
                        "type": "integer"
                      },
                      "currentPageNumber": {
                        "type": "integer"
                      },
                      "sortOrder": {
                        "type": "string"
                      },
                      "filterBy": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "paginationData": {
                  "itemsPerPage": 20,
                  "currentPageNumber": 0,
                  "sortOrder": "desc",
                  "filterBy": ""
                }
              }
            }
          }
        }
      }
    },
    "/notifications/mark-read/{id}": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "PATCH /notifications/mark-read/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Notification marked as read"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/notifications/mark-all-read": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "PATCH /notifications/mark-all-read",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "All notifications marked as read"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/notifications/delete/{id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "DELETE /notifications/delete/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Notification deleted successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/notifications/unread-count": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "POST /notifications/unread-count",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Unread count retrieved successfully",
                  "data": {
                    "unreadCount": 3
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
...
  }
}